requirements.txt
streamlit
python-dotenv
pytesseract
pdf2image
pillow
opencv-python
numpy
pandas
requests
openpyxl